Cleans up base project and renames zipkin to brave, as that's the tracer
There are many java libraries that write Zipkin data and use B3 headers.
This renames the Brave project to Brave, as that's what its name is.

This also cleans up a number of build issues including some missing
git ignore accidentally checked in project files, and lack of a maven
wrapper.